batch_queries_statement_timeout in a helper #10321
Move the functionality of batch_queries_statement_timeout to a common
module that can be used both by User and Carto::User classes.
